src/main/java/com/publicuhc/pluginframework/routing/SubcommandLengthComparator.java
/*
 * SubcommandLengthComparator.java
 *
 * The MIT License (MIT)
 *
 * Copyright (c) 2014 Graham Howden <graham_howden1 at yahoo.co.uk>.
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 * THE SOFTWARE.
 */

package com.publicuhc.pluginframework.routing;

import java.util.Comparator;

class SubcommandLengthComparator implements Comparator<CommandRoute>
{

    private boolean mostApplicable;

    public SubcommandLengthComparator(boolean mostApplicable)
    {
        this.mostApplicable = mostApplicable;
    }

    @Override
    public int compare(CommandRoute route1, CommandRoute route2)
    {
        int r1 = route1.getStartsWith().length;
        int r2 = route2.getStartsWith().length;

        return mostApplicable ? Integer.compare(r2, r1) : Integer.compare(r1, r2);
    }
}
